[lib] Stop requesting device lists for users that don't have them
Summary:
issue: [[https://linear.app/comm/issue/ENG-8842/userinfoshandler-calls-getdevicelistsforusers-with-the-same-users | ENG-8842]]
When we launch CSAT, some users will not yet have device lists. We don't want to be requesting them every couple of seconds.

Test Plan:
Created a user registared with the keyserver but not with identity. Checked that their device list was being fetched multiple times, and non was returned by identity.
Cehcked that after this change the list was being requested only once.

Reviewers:
tomek, kamil,
Subscribers: