[identity] increase connect timeout to 60 seconds
Summary: the DynamoDB client is timing out if it can't initiate a socket connection in 3.1 seconds. this diff bumps the timeout to 60 seconds.

it's still unclear to me why the db client is experiencing a connect timeout after making hundreds of API calls. i would assume that it's just reusing connections from the underlying connection pool...

but this change resolved the error i was seeing and identity was able to respond to my getDeviceListsForUsers request successfully

Test Plan: applied change to staging, confirmed that i got a successful response when requesting device lists for 12,000 user IDs. request took 33 seconds to complete

Reviewers: bartek, ashoat, will

Subscribers: