[keyserver] remove unnecessary inviteResponder line in keyserver.js
Summary:
We're  planning on running the webapp, keyserver, and landing page in isolation in separate containers for our prod self-host setup.

This means we should remove this line `server.get('/invite/:secret', inviteResponder);`

Test Plan:
flow check. I believe we plan on testing this on prod and check if the invite links still work. We can fully remove if confirmed working

Reviewers:
ashoat

Subscribers: