[lib] Turn on notifs by default for thick threads
Summary:
When I first introduced this code, I used `defaultThreadSubscription`, thinking it corresponding to having notifs turned on.

After testing as part of [ENG-8426](https://linear.app/comm/issue/ENG-8426/make-sure-rawthreadinfos-with-ids-that-dont-have-or-character-dont), I found that thick threads were being created with notifs turned off. This made me realize I should be using `joinThreadSubscription` here to make sure that the notifs are turned on by default for thick threads.

Depends on D12824

Test Plan: I followed the test plan described in [this Linear comment](https://linear.app/comm/issue/ENG-8426/make-sure-rawthreadinfos-with-ids-that-dont-have-or-character-dont#comment-f32630d3). Before this diff, the thick thread was created in the Muted tab. After this diff, it now appears in the Home tab

Reviewers: tomek, kamil, marcin

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D12825