[native][web] Allow sending custom encrypted message via Tunnelbroker test interface
Summary: Allowing this interface to send arbitrary messages helped me test DM ops in [ENG-8426](https://linear.app/comm/issue/ENG-8426/make-sure-rawthreadinfos-with-ids-that-dont-have-or-character-dont). I figured this might be useful in the future, and potentially to other memebers of the team.

Depends on D12825

Test Plan: I followed the test plan described in [this Linear comment](https://linear.app/comm/issue/ENG-8426/make-sure-rawthreadinfos-with-ids-that-dont-have-or-character-dont#comment-f32630d3), which relies on these changes to allow manually sending encrypted messages between clients

Reviewers: tomek, kamil, marcin

Subscribers: