[lib] Make sure thick threads starts out with startReached
Summary:
In order to make sure that `startReached` starts out as `true` when a thread is created, we set `truncationStatus` to `"exhaustive"`.

I thought about always using `exhaustive`, but this conflicts with our plans in the [Limit number of displayed DMs](https://linear.app/comm/project/limit-number-of-displayed-dms-ac87739cedfe) project, where we'll want `startReached` to be `false` if there are messages in SQLite that are not represented in Redux.

Depends on D12827

Test Plan: I followed the test plan described in [this Linear comment](https://linear.app/comm/issue/ENG-8426/make-sure-rawthreadinfos-with-ids-that-dont-have-or-character-dont#comment-f32630d3). Before this diff, the spinner was visible in the newly-created thick thread. After this diff, it was no longer visible

Reviewers: tomek, kamil, marcin

Subscribers: