[lib][native][web] Introduce threadTypeIsSidebar
Summary: In this diff we replace `threadTypes.SIDEBAR` checks with `threadTypeIsSidebar`, a new utility function that also checks `threadTypes.THICK_SIDEBAR`. This allows our business logic to treat the two in the same way.

Test Plan: I followed the test plan described in [this Linear comment](https://linear.app/comm/issue/ENG-8426/make-sure-rawthreadinfos-with-ids-that-dont-have-or-character-dont#comment-f32630d3), specifically for sidebar creation

Reviewers: tomek, kamil, marcin

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D12839