[identity] check users table before reserved usernames table for user ID during wallet login
Summary: we should check the users table first in case the user ID happens to be in both tables for some reason (this should never happen, but unfortunately there was a bug that led to this rule being violated). by checking the users table first, we ensure that registered wallet users will be able to log in even if there is invalid data in the reserved usernames table.

Test Plan: add userID and wallet address to staging reserved usernames table. then successfully log in with wallet

Reviewers: will, bartek

Subscribers: