[services] Replace deprecated aws_config::from_env()
Summary:
The `aws_config::from_env()` was deprecated in favor of `defaults()` which accepts a `BehaviorVersion` param.
I set the behavior explicitly to the latest available. We could also use `latest()` but I wanted to avoid it in case of breaking changes.

Depends on D12879

Test Plan: Made sure that services talk to localstack

Reviewers: varun!, will

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12880