[native] refresh device token after receiving message from Tunnelbroker
Summary:
[ENG-8498](https://linear.app/comm/issue/ENG-8498/[july]-handle-invalid-device-token).

This is similar to the keyserver's `BAD_DEVICE_TOKEN` logic.

This is only about `localToken` which is set for every user, not touching `tunnelbrokerToken` so it shouldn't affect devices without CSAT and TB connection.

Depends on D12912

Test Plan: Make Tunnelbroker throw an error with an invalid device token, and test that a new token is uploaded (value in DDB changed). Additionally verified this using logs.

Reviewers: bartek, inka, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12916