[native] linked devices component
Summary:
Depends on D12884

UI for linked devices. You can navigate from this screen to the SecondaryDeviceQRCodeScanner using the Add button on the top right

next diff will add a bottom drawer for interacting with each device

Test Plan:
successfully added a new device, it appeared in the list. navigation works as expected

{F2349070}

Reviewers: tomek, bartek, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12907