[lib] Stop requiring sequential messages
Summary:
[ENG-8715](https://linear.app/comm/issue/ENG-8715/stop-requiring-sequential-messages).

This differential is doing two things:
1. Renaming `decryptSequentialAndPersist` -> `decryptAndPersist` in `OlmAPI` to properly reflect what this code method is now doing.
2. Using `decrypt` instead of `decrypt_sequential` from `olm` which is an actual change, stops requiring sequential messages.

This might cause some messages to arrive in an inverted order, and we're planning to handle this here: https://linear.app/comm/issue/ENG-8752/update-the-reducers-to-respect-properties-timestamps

NOTE: This change is not revertable (in logic meaning), in the future if we want to start requiring sequential messages we'll need to reset the `olm` session.

Test Plan: This is just swapping back to the `decrypt` method which is used elsewhere in the codebase and has the same params and return type so this should work out of the box. Under the hood, we only avoid one additional check ([src](https://github.com/CommE2E/olm/blob/main/src/ratchet.cpp#L564-L567)).

Reviewers: marcin, tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12749