[native] Remove irrelevant code from nux tips overlay
Summary:
issue: [[https://linear.app/comm/issue/ENG-8882/scope-nux-further | ENG-8882]]
Removing code that is irrelevant to nux overlay.
Mind that this diff is only supposed to be removing unnecessary code. Name changes and so on will follow in the next diffs

Test Plan: Tested by creating a component wuth this `createTooltip`, and navigating to it. It appears correctly.

Reviewers: tomek, will

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12942