[lib] Apply the queued updates
Summary:
When a thread appears in the store, we should process all the operations that were queued for it.

https://linear.app/comm/issue/ENG-8769/apply-the-queued-updates

Depends on D12954

Test Plan: Processed an operation that updates thread name, with a thread ID that isn't present in the store. Then processed another operation that creates a thread with this ID. Checked that the name of the created thread matched the value from the first operation. Also checked that the queue no longer contains the operations.

Reviewers: kamil, inka

Reviewed By: inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12956