[native] Add nux context
Summary:
issue: [[https://linear.app/comm/issue/ENG-8618/find-coordinates-and-verticalbounds-for-tooltip | ENG-8618]]
To be able to get coordinest of buttons, we need to let them render first. Then, they will register their coordinates in this context. A handler component will select those coordinates, and navigate to screens passing the coordinates in params
Test Plan:
Tested with following diffs. Testeds that after a button registares their coordinates the handler effect is fired and nux tip appears on the screen, rendering the button in the cprrect place
Reviewers:
Tomek, will
Subscribers: