[native] Register CommunityDrawerButton in nux context
Summary: issue: [[https://linear.app/comm/issue/ENG-8618/find-coordinates-and-verticalbounds-for-tooltip | ENG-8618]]

Test Plan:
Tested with following diffs. Tested that after a button registeres their coordinates the handler effect is fired and nux tip appears on the screen, rendering the button in the correct place
Tested that as the button unmounts, the cleanup function is called. Checked that it is not called in unexpected moments

Reviewers: tomek, will

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12964