[lib] Correctly handle generating ops from multiple updates
Summary:
We have to apply the ops from each update before generating ops for the next update.

https://linear.app/comm/issue/ENG-8932/thread-updates-are-incorrectly-reduced

Depends on D12956

Test Plan: Modified the code so that a `change_thread_settings` operation, that updates name, color and unread status of a thread, is processed. Made sure that the changes were correctly applied on web.

Reviewers: kamil, inka

Reviewed By: inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12957