[lib] Add action type to remove peer user
Summary:
Similiar to D12961, but for `relyingOnAuthoritativeKeyserve = false`.
Added new action type that removes user IDs from both userStore and auxUserStore.
The action is dispatched when a peer receives message about account deletion.

Depends on D12961

Test Plan: The same as D12961, but this time faked `relyingOnAuthoritativeKeyserver = false` by reversing conditions for this flag in reducers.

Reviewers: inka, kamil, tomek

Reviewed By: inka, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12962