[terraform] Adds aws region export to aws-deploy.sh script
Summary:
Ashoat ran into an issue of having a different aws region configured through aws-cli when running aws-deploy which prevented the migrations from running.

We should ensure the script properly sets the region so the user will always be able to run the script

Test Plan:
I deployed my self-host setup to us-east-1. Through aws configure, I set the region to us-east-2 and ensured the aws-cli region was set by using aws-cli to
describe the keyserver cluster which resulted in a non-existent cluster. However, the script works properly, demonstrating that the script was able to properly
set the region to us-east-1

Reviewers: bartek, varun

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12959