[terraform] configure personal ip address through script and not allowed_ips var
Summary:
We rethought the approach in allowing script access to receive a health check from the keyserver load balancer.

Instead of allowing `allowed_ips` constant access to the load balancer, we decided to simply have the script enable access to the load balancer from the ip address of the
script caller by adding an ingress rule through aws cli during the duration of the script's run.

Test Plan: Ran the aws-deploy script and ensured that a separate ip machine was unable to access the load balancer while the script was able to perform a health check

Reviewers: varun, bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12960