Fix broken keyserver-based notifs when shared worker is behind service worker
Summary: This differential fixes https://linear.app/comm/issue/ENG-8987/web-notifs-became-broken-on-prod. There is no need to throw if there is no notiufs account in IndexedDB when decrypting keyserver notif since it is not necessary to potentially create session. Peer-based notifs have other checks in the code that will throw if there is no notifs account in indexeddb.

Test Plan:
1. Before this diff remove `notificationAccount` entry in IndexedDB. Send notif. Expect broken one.
2. Apply this diff and remove `notificationAccount` entry in IndexedDB. Send notif. Expect correct one.

Reviewers: tomek, kamil, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13009