[native] make SecondaryDeviceQRCodeScanner dismissible
Summary: If you reject the Alert that pops up when you scan a QR code, the same alert will pop up if your phone camera is still pointed at the QR code. Instead, we should just go back when the cancel button is pressed.

https://linear.app/comm/issue/ENG-8935/cant-dismiss-secondarydeviceqrcodescanner

Test Plan: confirmed that if I press cancel on the Alert, I'm navigated back to the linked devices screen

Reviewers: ashoat!

Subscribers: