[native] Fix VideoPlaybackModal for encrypted keyserver-hosted media
Summary:
Encrypted keyserver-hosted encrypted media works on web, but on native it looks like only blob-hosted encrypted media works.

I think this work was missed. I found we're using `encryptedMediaBlobURI` on web (and elsewhere on native), so I integrated it here and it looks like it did the trick.

Test Plan:
Prior to this, when opening keyserver-hosted encrypted media, the app would basically freeze... it would be stuck loading the content, but it wouldn't show anything.

After this diff, I confirmed that keyserver-hosted media loads correctly, and can be viewed.

Reviewers: bartek

Subscribers: