[lib] Don't check relyingOnAuthoritativeKeyserver when reducing removePeerUsersActionType
Summary:
There was some ambiguous feedback in D12962 – @bartek interpreted that this check should be left, but I interpreted that it should be removed.

My intuition is that we should reduce `UserStore` and `AuxUserStore` under the same conditions. Curious for other reviewers' take.

Test Plan: I haven't tested this. If reviewers think it needs to be tested, I would appreciate @bartek's help on that, as he wrote the initial diff and has context on how to test

Reviewers: inka, kamil, bartek

Reviewed By: kamil

Subscribers: tomek, bartek

Differential Revision: https://phab.comm.dev/D13041