Extract logic of pushInfo -> usersToCollapsableNotifInfo to lib and use for client-based notifs
Summary: This differential extracts teh logic that groups notifications message infos by collapse key to lib. This is necessary to have working e2ee notifs for sidebar creation. Without this work we would call `notificationsTexts` on `CREATE_SIDEBAR` message data and this crashes the app.

Test Plan: For now test that keyserver notifs are not regressed by this diff.

Reviewers: tomek, kamil

Subscribers: