[lib] Allow sending messages to some users
Summary:
We need this because we plan to send a different message to users who are added to a thread and to users who already were members, when we add new members.

https://linear.app/comm/issue/ENG-8929/split-add-members-operation

Depends on D13053

Test Plan:
Checked two scenarios:
Open two clients with two different users logged in. Send two messages, one with the user included and one without. Check if only one message was delivered.

Reviewers: kamil, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13055