[native] Simplify types
Summary:
issue: [[https://linear.app/comm/issue/ENG-8914/refactor-nux-overlay | ENG-9814]]
I don't need to pass progress to the button components. This is needed for example in text message so that it can move to the right space after being pressed (in oreder to display tooltip). The nux tips don't need this

Test Plan: N/A - this code isn't being used anywhere yet, and the change is very simple

Reviewers: tomek, will

Reviewed By: tomek, will

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13065