[native] Add mappaing from nux tips to button onPress params
Summary:
issue: [[https://linear.app/comm/issue/ENG-8617/create-nux-tooltip-item | ENG-8617]]
A handler will display one tip, then pressing on `ok` will navigate to the next tip and so on. So we need to be able to provide each tip with coordinates and params for the next tip.
We will do this like so: route.params will include a `tipKey`, and the tip will fetch its coordinates from the context. It will also use this `getTipCallbackParams` function to get params for the next tip it has to navigate to.

Test Plan: Tested with later diffs - tested that this solution works and it is possible to navigate to next tips

Reviewers: tomek, will

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13077