[terraform] remove unnecessary aws deploy script commands to adjust ingress rules to keyserver load balancer
Summary:
Running the function to restore ingress rules is unnecessary as running terraform apply already takes care of this. The script is not broken with this function, but will
output unnecessary output noting that authorizing the general ip access cidr of 0.0.0.0/0 already exists and there is no personal ip ingress rule to unauthorize.

Test Plan: Tested by running script and was able to access the keyserver

Reviewers: bartek, varun, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13087