[terraform] increase max time for curl
Summary:
On running aws-deploy script, we ran into the issue of 5 seconds being too small and causing
the script to exit early when a curl was still possible when set to 15 seconds.

Depends on D13093

Test Plan:
ran the script. Behavior did not change. Unsure of how to replicate initial
conditions which required 15 seconds

Reviewers:
ashoat,
varun,
bartek

Subscribers: