[terraform] remove unnecessary redeploy
Summary:
These parts of the script are no longer necessary as they are taken care of by
the terraform apply. Originally had the redeployment to take care of redeploying
the primary service after all nodes were taken down, but terraform apply already restores 1 desired node for primary.

Additionally, when the image is updated to a new one (which currently is assumed to cause a migration), local provisioner logic
in the terraform apply will ensure that the primary service is available before the secondary service is launched

Depends on D13094

Test Plan:
ran the script and confirmed migration occurred successfully and secondary nodes only applied after primary node was healthy

Reviewers:
varun,
bartek

Subscribers: