[terraform] run migration on aws deploy script only when necessary
Summary:
This includes changes to the script that would allow it to only run the migration
when there is an upgrade to the database version. If the old and intended db versions are the same, a terraform apply occurs.

It fails when the db version is intended to be downgraded. When db version is not specified as a label,
the script will ask the user if they would like the force an upgrade as the intended db version is unknown.

Test Plan:
Ran this starting with a keyserver running db version 66 on my own custom image. Tested by upgrading to another custom image
running on db version 68. This change was noted by the script and it prompted the user if they would like to go forward with the migration.

Trying to go backwards was aborted. And trying the commapp/keyserver:1.0.105 resulted in the prompt asking if the user would like
to force an image upgrade which results in the complete migration process.

Reviewers: varun, bartek, ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13104