[lib] Propagate MEMBER permissions to all chats in a community
Summary:
This addresses [ENG-9054](https://linear.app/comm/issue/ENG-9054/member-doesnt-have-permission-to-react-if-they-arent-a-member-of).

The issue is that in D13014, I return `false` from `includeThreadPermissionForThreadType` even if a propagation prefix is set.

We want role permissions to apply to all chats in a community, even if there is a "broken link" where you are a member of a grandchild but not a member of its parent.

Depends on D13112

Test Plan:
1. I set up a repro of the issue in my local environment
2. I applied this diff to fix the issue
3. I reset `db_version` to 67 to trigger migration 68 again
4. After migration 68 ran again, I confirmed that the issue was gone

Reviewers: tomek, inka

Subscribers: