[CI] Introduce GitHub Actions CI job for keyserver releases
Summary:
This addresses [ENG-9048](https://linear.app/comm/issue/ENG-9048/github-action-to-generate-keyserver-docker-image).

We'll use the same image for keyserver, web, and landing.

For now, I'll continue to handle the creation of entries in the GitHub Releases page manually. This is how we do it for most workflows, with the exception of `desktop_release.yml`.

@will will follow-up with some changes here as part of [ENG-8839](https://linear.app/comm/issue/ENG-8839/determine-when-migrations-are-necessary-in-aws-deploysh).

Test Plan:
I'm not sure how to test this yet. I could land it and try testing from the master branch directly. From my research, it seems like the main alternative would be to test from a fork, but that would rewuire duplicating all relevant GitHub Secrets there

Reviewers: will

Subscribers: