[lib] Make FarcasterDataHandler go through useFindUserIdentities
Summary:
There are currently two components that call `findUserIdentities`: `UserInfosHandler` and `FarcasterDataHandler`.

The former goes through `useFindUserIdentities`, but the latter accesses `findUserIdentities` directly via the `IdentityClientContext` context.

This diff makes both of them go through `useFindUserIdentities`, as preparation for adding a caching layer to that hook.

This is the first step to addressing [ENG-9002](https://linear.app/comm/issue/ENG-9002/make-sure-client-doesnt-query-finduseridentities-multiple-times-for).

Test Plan: This is a simple refactor. I made sure that `IdentityClientContext` should always be set where `FarcasterDataHandler` is rendered on both native and web.

Reviewers: varun, will

Reviewed By: varun

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13117